Define package, all grovel types; load libffi; define structs
Define the :fsbv package.  Define types that are defined in the
library.  Define structs: ffi-type (ffi_type) is defined using the
groveller because it has a name after the "typedef struct" in ffi.h;
ffi-cif (ffi_cif) is defined by hand in structs.lisp because it
doesn't, so gcc can't handle it.  Everything compiles and load in
linux/amd64.