New macro defcstruct
Introduce new macro fsbv:defcstruct which expands to cffi:defcstruct
and a form that bind the symbol property 'type-pointer to the libffi
structure.  This means that to define a new structure that can be
passed to and returned from a foreign function, the user need only
declare with fsbv:defcstruct, whose syntax is identical to
cffi:defcstruct.  The example structure 'complex is now defined this
way, and the example functions work.  Several very small files have
now been consolidated into two files init.lisp and defs.lisp.