Args :constructor, :deconstructor to defcstruct name-and-options
In defcstruct, you may supply :constructor (default 'list) and
:deconstructor (default 'elt) that will assemble or disassemble a CL
object from the pieces in the foreign struct, respectively.  The
deconstructor can be a symbol (as in the default value), in which case
each component is extracted by calling that function with the object
and then a sequence number, e.g. (elt foreign-struct 2), or it may be
a list of symbols, in which case each component is extracted by apply
each function in succession.  The example of complex numbers has been
updated.