Macro with-foreign-objects, functions object and (setf object)
The macro with-foreign-objects works like CFFI's except that it takes
an optional initial-value argument, which it uses to initialize the
foreign object.  The functions object and (setf object) (renamed from
foreign-object-components) gets and sets respectively the foreign
object to/from a Lisp object.  The example and readme.html have been
updated to reflect the current state of the code.