Use structure slot name from defcstruct
In expansion for 'foreign-object-components and
'setf-foreign-object-components, read the structure slot name from the
defcstruct, rather than the hardwired structure slot name 'dat which
was left over from the complex number example.  There is a potential
inefficiency now in the expansion because the pointer to the first
element of the filed is recomputed for each one; comment added to the
top of the file.