New function cstruct-libffi-type-pointer
Created a function cstruct-libffi-type-pointer that generates the
libffi-type structure for any foreign structure type, so that this
structure may be used as an argument or return call by value to a
foreign function.  It replaces the automatic generation of this
structure when defstruct is compiled or loaded (via
*defcstruct-hook*), so that it is only made when needed and cffi-fsbv
need not be loaded when the defcstruct is compiled.  The structure
will only be made once; it is bound to the plist of the type symbol
(as before), and is reused if needed again.  This function compiles
but has not yet been tested.