A pointer to a struct is an aggregate type
The deprecated bare struct specification (meaning a pointer to a
struct) returned T for aggregatep, but the new specification (:pointer
(:struct foo)) returned NIL, because a pointer is a built-in type.
This caused an error in mem-aref pointed out by Ryan Pavlik.  Added
his example cases.  There is a difference in pointer value between the
deprecated form and the new form when the offset is not 0.