Fix Solaris compilation options.
The "-fPIC" flag will always be necessary when using GCC-derived compilers as I understand it.

Additionally, Solaris compilation options now unconditionally tries a
64bit build.  Contemporary Solaris (since Solaris 10), are hybrid
32/64 bit environments where "uname -m" reports "i86pc".  Whether one
wants the 32 or 64 bit version of the libraries depends on the type of
the Lisp implementation doing the loading, which is currently not
available in the environment in which make(1) is executing.