Indirect string one level for libffi
Every argument to libffi is dereferenced before being passed to the
function, including pointers. For strings to be correctly passed, an
additional level of indirection is needed. This fixes the need to use
the "address of" operator `&' in the C function for strings passed to
foreign functions that also take structures by value.