Restructure foreign-funcall-form and ffcall-body-libffi
Restructure foreign-funcall-form and ffcall-body-libffi so that the
fsbvp split is made first in the former, and the translate-objects is
called separately for the cases non-fsbvp, and fsbvp in the latter.
This should result in the same macroexpansion, but facilitates further
changes that will fix the macroexpansion so that translation of the
foreign object occurs before exiting the with-foreign-objects that
created it.