New #'translate-objects-ret translates return type
New function translate-objects-ret translates the return type even
when it's of a built-in type. Previously, this was done inside the
translate-objects call in ffcall-body-libffi. This worked because
translate-objects didn't touch the built-in object, but this is a
little easier to follow, because translate-objects-ret wraps a call to
translate-objects with a (possible) translation of the return object.