Make CFFI-grovel play well with ASDF 3.1.6
Move the C toolchain support to its own system cffi-toolchain.

Create linkable object files and include them in output-files for
compile-op, so that they can be found and linked by a recent-enough
ASDF (3.1.6 or later).

Allow for delivery of a system with a single static and/or dynamic
library, and on support platform (CLISP, CMUCL, SBCL), enable delivery
of a standalone executable that statically links required extensions.

This requires a recent SBCL (will hopefully be released in 1.2.17).
ASDF 3.1.6 is required, but backward compatibility code is
provided at least back to ASDF 3.1.2.

On BSD, use ld -r for "libraries" instead of ar and possibly libtool.

When forking a Lisp to dump an image, support Quicklisp.

Make CFFI-Grovel generally more robust.

Document static linking operations in manual. Improve documentation.

Add test for ASDF support including wrappers.