Fix libffi with enum argument
The function #'expand-to-foreign-dyn-indirect, used by cffi-libffi, did
not create the correct expansion when there was an enumerated argument,
because two translations are required, one from the symbol to an
integer, and the second to make the integer indirect, passing a pointer
to libffi. This remedies that problem by defining a method for
expand-to-foreign-dyn-indirect specific to foreign-enum that recursively
calls expand-to-foreign-dyn-indirect twice. A test, fsbv.8, has been
created.