Fixup usage of CFLAGS and CC environment variable
CFLAGS was not used at all, and CC was read only at load time, which
could be problematic if e.g. an image were saved with cffi-toolchain
already loaded, then changes to the environment variables would have no
effect.

The new behavior for both variables is as follows:

1) If the user has modified (either set or bound) the special variable,
   then use that value.
2) Otherwise use the environment variable, if present.
3) Otherwise use the default value.

There is also a new function provided: cffi-toolchain:cc-flags; this
performs the above algorithm.  This is used by the groveller to pick up
on any changes to the environment variable since load-time, and can be
used by others if they wish to e.g. add new values to *cc-flags*.

I would have liked to use a symbol-macro to remove the need for a
function, but dynamic bindings and symbol-macros don't mix; dynamically
binding *cc-flags* is a common use-case.