system: re-add dependency on UIOP
Dependency on `uiop' can't be implicit for sake of `program-op' and
`image-op' builds. User may wish to not include `asdf' or `cmp' (in case
of ECL) in the final image. Also not including this dependency breaks
the build-system and the application separation and is clearly invalid
from the system definition point of view.

ASDF has option (require-system :system) which doesn't reload it, if
it's already present in the image (even if it's newer):

https://common-lisp.net/project/asdf/asdf.html#index-require_002dsystem

so if we create alternative patch like:

-  :depends-on (:alexandria :trivial-features :babel)
+  :depends-on ((:require :uiop) :alexandria :trivial-features :babel)

then ASDF shouldn't reload the system if it's present in the memory.

That said, ASDF doesn't handler (:require :system-name) options
correctly yet.

(see https://gitlab.common-lisp.net/asdf/asdf/merge_requests/13).

Closes #96.