Add ffi-form-generators feature to c2ffi
The goal of this feature is to enable people to emit their own code
based on the generated form.

To use this you specify one or more generator forms using the
`:ffi-form-generators` key in your `:cffi/c2ffi-file` component.

For example:

    :ffi-form-generators ((:defcstruct "raw-bindings-nuklear.ffi::ffi-generator-struct"))

This means that for every `cffi:defcstruct` form that is written to the
generated file `raw-bindings-nuklear.ffi::ffi-generator-struct` will be
called. Whatever form is returned from that function will be written
into the generated file using `#'output/code`.

The first element of the form (:defcstruct in the above case) is called
the 'pattern' and the second is the 'generator function'.

You can define multiple generator forms:

    :ffi-form-generators ((:defcstruct "raw-bindings-nuklear.ffi::ffi-generator-struct")
                          (:defcunion "raw-bindings-nuklear.ffi::ffi-generator-union"))

It is legal to have both call the same function if you like.

These are the valid patterns you can use in your generator forms:

    :defcfun
    :defcstruct
    :defcunion
    :defctype
    :defcvar
    :defbitfield
    :defcenum
    :begin
    :end

A generator function using the `:begin` pattern will be called once,
just before the c2ffi entries are processed

A generator function using the `:begin` pattern will be called once,
just after the c2ffi entries are processed

The signature for your generator function should be as follows:

    (defun your-function (definition context &key &allow-other-keys)
      ...)

The context argument will be an instance of the `ffi-generator-context`
class. This object has one field called data, it has an accessor called
`#'context-data`.

The context exists for cases where you want to aggregate data during the
processing, usually with the goal of emitting it at the `:end`.

This can be useful when you want to make sure you don't emit forms for
duplicate definitions of cffi types. For example you may want to emit an
accessor function for cetain CFFI structs, however SBCL will throw an
error if it finds duplicate `defun`s in the same file

Here is an example of how this feature can be used. The goal is to emit
a `defstruct` form for every `cffi:defcstruct`.

In the asd file:

    :ffi-form-generators ((:begin "raw-bindings-nuklear.ffi::ffi-generator-begin")
                      (:defcstruct "raw-bindings-nuklear.ffi::ffi-generator-struct")
                      (:end "raw-bindings-nuklear.ffi::ffi-generator-end"))

In a auxilary lisp file:

    ;; we make a hash table to store and deduplicate the defstruct forms
    (defun ffi-generator-begin (definition context &key &allow-other-keys)
      (declare (ignore definition))
      (setf (context-data context) (make-hash-table :test #'equal))
      nil)

    ;; For each defcstruct we make a defstruct form and store it in the
    ;; context
    (defun ffi-generator-struct (definition context &key &allow-other-keys)
      (destructuring-bind (name-and-options . fields) (rest definition)
        (destructuring-bind (name . options) (uiop:ensure-list name-and-options)
          (declare (ignore options))
          (when fields
            (let* ((wrapper-name-str (format nil "~a-WRAPPER" name))
                   (wrapper-name (intern wrapper-name-str)))
              (setf (gethash `(:struct ,name) (context-data context))
                    `(defstruct ,wrapper-name ,@(mapcar #'first fields)))))))
      nil)

    ;; We emit all the defstruct forms
    (defun ffi-generator-end (definition context &key &allow-other-keys)
      (declare (ignore definition))
      (let ((data (context-data context)))
        `(progn
           ,@(loop :for k :being :the :hash-keys :of data :collect
                (gethash k data)))))

This feature is slightly more open ended than the other :ffi-*
transformer and export forms, however I feels it's flexibility worth the
small cost in difficulty.