Make it possible to customize c2ffi form generation
The goal of this feature is to enable people to emit their own code
based on the generated form.

The caller can specify a plugin factory function that can return up to
2 functions using VALUES.

The first function is called with every definition that is emitted. The
return value of this function is added to the output file using
OUTPUT/CODE. NIL is a valid return value.

The second function is called once after all the definitions have been
emitted. No arguments are passed. The result should be NIL or a list of
forms to be emitted as if by `(map nil 'output/code result-list)`.