Rewrite ffi-form-generator feature
The ffi-form-generator argument allows you to specify a function to be
called once for every form that is written into the generated lisp file.

The generator function is specified like this:

    :ffi-form-generator "function-name"

any non nil value returned from the generator function will be written
into the generated lisp file with #'output/code.

An example function would be:

    (defun ffi-form-generator (definition &key &allow-other-keys)
      (let ((def-name (second definition)))
        (when (eq def-name 'defcstruct)
          (let* ((name (intern (format nil "~a-WRAPPER" def-name))))
            `(defstruct ,name (mapcar #'first (cddr definition)))))))

This will emit one lisp wrapper struct for every C Struct emitted.

You can optionally also return a second value that will be used as a
deduplication-key. CFFI will collect all of the forms returned with a
deduplication-key, remove the duplicates, and emit all the forms at the
end of the generated lisp file.

For example if we decided that we want to make sure that our earlier
example does not emit multiple defstructs if there are multiple C
structs we can do the following.

    (defun ffi-form-generator (definition &key &allow-other-keys)
      (let ((def-name (second definition)))
        (when (eq def-name 'defcstruct)
          (let* ((name (intern (format nil "~a-WRAPPER" def-name))))
            (values `(defstruct ,name (mapcar #'first (cddr definition)))
                    def-name)))))