switch to using :ffi-plugin-factory
This time we allow the user to specify a plugin factory function.

The factory function can return up to 2 functions using `#'values`.

The first function is called with every definition that is emitted. The
return value of this function is added to the output file using
`#'output/code`. nil is a valid return value

The second function is called once after all the definitions have been
emitted. No arguments are passed. The result should be nil or a list of
forms to be emitted as if by `(map nil #'output/code result-list)`.