Null name shouldnt satisfy strong rules
The problem with this that with forms like this:

    typedef enum {
      dJointTypeNone = 0,
      dJointTypeBall,
      dJointTypeHinge,
    } dJointType;

The enum is always excluded, as the name is nil. Then the `dJointType`
typedef tries too lookup the anonomous entry for the enum and fails.

The only way to counter it is to have explicitly include the definitions
for everything in the files (which makes it feel much less automatic).

This commit is one possible fix. Another is to allow the user to
customize the `include-definition?` predicate. Also `cl-autowrap`
handles this case so maybe we can take some inspiration from there.