Make static linking command more robust
Ensure symbols are printed and read from package asdf, instead of printed
from package common-lisp and read from what is probably common-lisp-user
but without 100% guarantee. This also makes the printed output easy to
copyable and pastable into any syntactic context that will let you use
require and in-package. Finally symbols in sub-packages of ASDF won't be
printed as being of their particular subpackage, which works better if there
are package discrepancies between the implementation-provided ASDF and the
configured ASDF to which it will be upgraded at the first command.

Add some comments and reformat some code.