Test DEFWRAPPER with void returns and pointers
Make sure that DEFWRAPPER doesn't insert a "return <foo>" statement when
wrapping a macro/function that returns void, and confirm that the Lisp code
for binding to C is properly generated.

Because string comparisons are used to check the generated C code, some
non-critical things like the name mangling or formatting of the code can
affect whether the test passes or fails. This means that if the C code
generated by DEFWRAPPER changes, these tests will need to be changed.
Even if the new C code is semantically equivilant.