toolchain: remove non-system libaries in STATIC-IMAGE-OP
Which applies to STATIC-PROGRAM-OP as well, because it's a
subclass. Essentially, the non-system libraries have been statically
linked (C-wise) into the image, so we don't need to have them as
foreign libraries. In fact, as foreign libraries, the restored image
will try to load them and fail, hence defeating the point.

System libraries do need to be kept, though, as they're shared
libraries, and not embedded in the image at all.