Register the c-file libraries as :static.
Note, however, that it is not running LOAD-FOREIGN-LIBRARY, as that
would have it reloaded when the image is restored. For example, SBCL
stores the loaded foreign libraries in its "shared objects" concept,
and tries to restore them all by default when restoring. See also
https://github.com/cffi/cffi/pull/157/ for a similar effect.

By registering the c-file libraries as :static, using the
recently-merged FOREIGN-LIBRARY-LOAD-STATE, we can make CFFI not use
dlopen() after the image is restored, when the library will be needed.