abcl: implement MAKE-SHAREABLE-BYTE-VECTOR
In abcl-1.7.0, CL:MAKE-ARRAY is able create and share malloc()d memory
for java.nio.ByteVector allocated by JNA via use of the :NIO-BUFFER
argument. Whether this feature is present in the current
implementation is conditionalized on the presence :NIO in
CL:*FEATURES*.

We retain the old behavior of copying memory for
WITH-POINTER-TO-VECTOR-DATA when passed a non-shareable byte vector as
a plausible fallback strategy.