Fix macro discard-docstring
When the given list has one element, `(cdr list)` is nil, so it  must be
negated for the condition to ever be true in the correct
circumstances. This code appears to try to handle a situation in
`defcallback` that no longer occurs, so the condition is reverted back
into its original form.