cmucl vararg fix
https://gitlab.common-lisp.net/cmucl/cmucl/-/issues/74
led to this fix
https://github.com/cffi/cffi/pull/361
revision: c42b42e93a784c9396047078eff9caed257d7c28
but the fix was only partial.

This patch completes that fix.

Elaboration:

(defcfun ...)
expands into a defun if given fixed args or a defmacro if given varargs.
The original fix addressed the fixed-arg case but not the vararg case.

For the vararg case where defcfun expands into a defmacro, it's the
callers of that macro that need protecting with notinline, but they
don't get that protection.

One solution is to chase down the code for that other case and
add a notinline declaration to it. That leads to early-types.lisp's
expand-to-foreign-dyn :around method's changing from
(unwind-protect (progn ,@body))
to
(unwind-protect
  ((lambda ()
     #+cmucl (declare (notinline alien::%heap-alien))
     ,@body)))
but I don't know if there are other possible places that should also be
modified to have a notinline declaration.

A better solution regardless is move the existing notinline declaration
deeper to a single place from which both cases get protected. That's
what this patch does.

The following snippet fails without this patch and passes with it.
(cffi:defcfun ("H1" h2) :void &rest)
(defun h () (h2))
(compile 'h)